from_json without module
