+++ src/daft-functions-json/Cargo.toml
@@ -6,6 +6,7 @@ jaq-core = {workspace = true}
 jaq-interpret = {workspace = true}
 jaq-parse = {workspace = true}
 jaq-std = {workspace = true}
+pyo3 = {workspace = true, optional = true}
 serde = {workspace = true}
 serde_json = {workspace = true}
 typetag = {workspace = true}
