+++ tests/io/json/assets/single.json
@@ -1,0 +1,4 @@
+{
+    "a": 1,
+    "b": 2
+}
